# spec/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Workflow Orchestration API
  version: 0.1.0
servers:
  - url: /api/v0
paths:
...
  /workflows/{workflowId}:
    get:
      summary: Obtener workflow por id
      parameters:
        - in: path
          name: workflowId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
        '404':
          description: Not Found
    put:
      summary: Actualizar workflow (solo draft)
      responses:
        '200':
          description: OK
        '409':
          description: Conflict
...
  /runs/{runId}:
    get:
      summary: Obtener ejecución
      parameters:
        - in: path
          name: runId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
        '404':
          description: Not Found
...